Revert and create new file
