+++ packages/manager/src/components/GravatarByEmail.tsx
@@ -3,8 +3,6 @@ import { styled } from '@mui/material/styles';
 import * as React from 'react';
 
 import UserIcon from 'src/assets/icons/account.svg';
-import ProxyUserIcon from 'src/assets/icons/parent-child.svg';
-import { Box } from 'src/components/Box';
 import { getGravatarUrl } from 'src/utilities/gravatar';
 
 interface Props {
@@ -35,46 +33,4 @@ const StyledAvatar = styled(Avatar, {
 })<Partial<Props>>(({ height, width }) => ({
   height: height || 28,
   width: width || 28,
-}));
-
-type ProxyUserGravatarProps = Omit<Props, 'email'>;
-
-export const ProxyUserGravatar = ({
-  height = 34,
-  width = 34,
-}: ProxyUserGravatarProps) => {
-  return (
-    <Box
-      sx={(theme) => ({
-        background: `linear-gradient(60deg, ${theme.color.blue}, ${theme.color.teal} )`,
-        backgroundSize: '300% 300%',
-        borderRadius: '50%',
-        height,
-        padding: '3px',
-        width,
-      })}
-    >
-      <Box
-        sx={(theme) => ({
-          background: theme.color.white,
-          borderRadius: '50%',
-          color: theme.palette.text.primary,
-          height: `calc(${height}px - 6px)`,
-          overflow: 'hidden',
-          position: 'relative',
-          width: `calc(${width}px - 6px)`,
-        })}
-      >
-        <StyledProxyUserIcon />
-      </Box>
-    </Box>
-  );
-};
-
-const StyledProxyUserIcon = styled(ProxyUserIcon, {
-  label: 'styledProxyUserIcon',
-})(() => ({
-  bottom: 0,
-  left: 0,
-  position: 'absolute',
 }));
